feat: put api in a new repo of its own
The goal of the file sharing web app was to learn angular. It seems best to
separate the front end and the back end into their own repositories due to
how cluttered the original workspace had become. Originally I was only trying
to learn angular with a vague concept of an app, making it up as I went, but
over time it grew into something I had use for. I didn't bother to make any
commits to the repo until after it took shape.